+++ Controller/PI/Callback3D.php
@@ -5,20 +5,21 @@ namespace Ebizmarts\SagePaySuite\Controller\PI;
 use Ebizmarts\SagePaySuite\Api\Data\PiRequestManagerFactory;
 use Ebizmarts\SagePaySuite\Model\Api\ApiException;
 use Ebizmarts\SagePaySuite\Model\Config;
+use Ebizmarts\SagePaySuite\Model\CryptAndCodeData;
+use Ebizmarts\SagePaySuite\Model\Logger\Logger;
 use Ebizmarts\SagePaySuite\Model\PiRequestManagement\ThreeDSecureCallbackManagement;
 use Ebizmarts\SagePaySuite\Model\RecoverCart;
 use Ebizmarts\SagePaySuite\Model\Session as SagePaySession;
 use Magento\Checkout\Model\Session as CheckoutSession;
 use Magento\Customer\Api\CustomerRepositoryInterface;
 use Magento\Customer\Model\Session as CustomerSession;
+use Ebizmarts\SagePaySuite\Helper\CustomerLogin;
 use Magento\Framework\App\Action\Action;
 use Magento\Framework\App\Action\Context;
+use Magento\Framework\App\CsrfAwareActionInterface;
 use Magento\Framework\App\Request\InvalidRequestException;
 use Magento\Framework\App\RequestInterface;
 use Magento\Sales\Api\OrderRepositoryInterface;
-use Magento\Framework\App\CsrfAwareActionInterface;
-use Ebizmarts\SagePaySuite\Model\CryptAndCodeData;
-use Ebizmarts\SagePaySuite\Model\Logger\Logger;
 
 class Callback3D extends Action implements CsrfAwareActionInterface
 {
@@ -52,6 +53,9 @@ class Callback3D extends Action implements CsrfAwareActionInterface
     /** @var CustomerRepositoryInterface */
     private $customerRepository;
 
+    /** @var CustomerLogin */
+    private $customerLogin;
+
     /**
      * Callback3D constructor.
      * @param Context $context
@@ -65,6 +69,7 @@ class Callback3D extends Action implements CsrfAwareActionInterface
      * @param Logger $suiteLogger
      * @param CustomerSession $customerSession
      * @param CustomerRepositoryInterface $customerRepository
+     * @param CustomerLogin $customerLogin
      */
     public function __construct(
         Context $context,
@@ -77,7 +82,8 @@ class Callback3D extends Action implements CsrfAwareActionInterface
         CheckoutSession $checkoutSession,
         Logger $suiteLogger,
         CustomerSession $customerSession,
-        CustomerRepositoryInterface $customerRepository
+        CustomerRepositoryInterface $customerRepository,
+        CustomerLogin $customerLogin
     ) {
         parent::__construct($context);
         $this->config = $config;
@@ -91,20 +97,24 @@ class Callback3D extends Action implements CsrfAwareActionInterface
         $this->suiteLogger                 = $suiteLogger;
         $this->customerSession             = $customerSession;
         $this->customerRepository          = $customerRepository;
+        $this->customerLogin               = $customerLogin;
     }
 
     public function execute()
     {
+        $encryptedOrderId = $this->getRequest()->getParam("orderId");
+        $orderId = $this->decodeAndDecrypt($encryptedOrderId);
         try {
             $sanitizedPares = $this->sanitizePares($this->getRequest()->getPost('PaRes'));
-            $encryptedOrderId = $this->getRequest()->getParam("orderId");
-            $orderId = $this->decodeAndDecrypt($encryptedOrderId);
             $order = $this->orderRepository->get($orderId);
             $customerId = $order->getCustomerId();
+
             if ($customerId != null) {
-                $this->logInCustomer($customerId);
+                $this->customerLogin->logInCustomer($customerId);
             }
+
             $payment = $order->getPayment();
+
             if ($this->isParesDuplicated($payment, $sanitizedPares)) {
                 $this->javascriptRedirect('checkout/onepage/success');
                 return;
@@ -125,6 +135,7 @@ class Callback3D extends Action implements CsrfAwareActionInterface
             $data->setVendorName($this->config->getVendorname());
             $data->setMode($this->config->getMode());
             $data->setPaymentAction($this->config->getSagepayPaymentAction());
+            $data->setSaveToken((bool)$this->getRequest()->getParam("saveToken"));
 
             $this->checkoutSession->setData(SagePaySession::PARES_SENT, $sanitizedPares);
 
@@ -139,19 +150,17 @@ class Callback3D extends Action implements CsrfAwareActionInterface
                 $this->javascriptRedirect('checkout/cart');
             }
         } catch (ApiException $apiException) {
-            $this->recoverCart->setShouldCancelOrder(true)->execute();
+            $this->recoverCart->setShouldCancelOrder(true)->setOrderId($orderId)->execute();
             $this->suiteLogger->sageLog(Logger::LOG_EXCEPTION, $apiException->getTraceAsString(), [__METHOD__, __LINE__]);
             $this->messageManager->addError($apiException->getUserMessage());
             $this->javascriptRedirect('checkout/cart');
         } catch (\RuntimeException $runtimeException) {
-            $orderId = $this->getRequest()->getParam("orderId");
-            $orderId = $this->decodeAndDecrypt($orderId);
             $vpstxid = $this->getRequest()->getParam("transactionId");
             $message = self::DUPLICATED_CALLBACK_ERROR_MESSAGE . ' OrderId: ' . $orderId . ' VPSTxId: ' . $vpstxid;
             $this->suiteLogger->sageLog(Logger::LOG_REQUEST, $message, [__METHOD__, __LINE__]);
             throw new \RuntimeException(__(self::DUPLICATED_CALLBACK_ERROR_MESSAGE));
         } catch (\Exception $e) {
-            $this->recoverCart->setShouldCancelOrder(true)->execute();
+            $this->recoverCart->setShouldCancelOrder(true)->setOrderId($orderId)->execute();
             $this->suiteLogger->sageLog(Logger::LOG_EXCEPTION, $e->getTraceAsString(), [__METHOD__, __LINE__]);
             $this->messageManager->addError(__("Something went wrong: %1", $e->getMessage()));
             $this->javascriptRedirect('checkout/cart');
@@ -223,18 +232,5 @@ class Callback3D extends Action implements CsrfAwareActionInterface
     public function decodeAndDecrypt($data)
     {
         return $this->cryptAndCode->decodeAndDecrypt($data);
-    }
-
-    /**
-     * @param $customerId
-     */
-    public function logInCustomer($customerId)
-    {
-        try {
-            $customer = $this->customerRepository->getById($customerId);
-            $this->customerSession->setCustomerDataAsLoggedIn($customer);
-        } catch (\Magento\Framework\Exception\LocalizedException $e) {
-            $this->suiteLogger->sageLog(Logger::LOG_EXCEPTION, $e->getTraceAsString(), [__METHOD__, __LINE__]);
-        }
     }
 }