+++ Controller/PI/Callback3Dv2.php
@@ -13,6 +13,7 @@ use Ebizmarts\SagePaySuite\Model\RecoverCart;
 use Magento\Checkout\Model\Session;
 use Magento\Customer\Api\CustomerRepositoryInterface;
 use Magento\Customer\Model\Session as CustomerSession;
+use Ebizmarts\SagePaySuite\Helper\CustomerLogin;
 use Magento\Framework\App\Action\Action;
 use Magento\Framework\App\Action\Context;
 use Magento\Framework\App\CsrfAwareActionInterface;
@@ -60,6 +61,9 @@ class Callback3Dv2 extends Action implements CsrfAwareActionInterface
     /** @var CustomerRepositoryInterface */
     private $customerRepository;
 
+    /** @var CustomerLogin */
+    private $customerLogin;
+
     /** @var Logger */
     private $suiteLogger;
 
@@ -78,6 +82,7 @@ class Callback3Dv2 extends Action implements CsrfAwareActionInterface
      * @param OrderLoader $orderLoader
      * @param CustomerSession $customerSession
      * @param CustomerRepositoryInterface $customerRepository
+     * @param CustomerLogin $customerLogin
      * @param Logger $suiteLogger
      */
     public function __construct(
@@ -94,6 +99,7 @@ class Callback3Dv2 extends Action implements CsrfAwareActionInterface
         OrderLoader $orderLoader,
         CustomerSession $customerSession,
         CustomerRepositoryInterface $customerRepository,
+        CustomerLogin $customerLogin,
         Logger $suiteLogger
     ) {
         parent::__construct($context);
@@ -110,21 +116,24 @@ class Callback3Dv2 extends Action implements CsrfAwareActionInterface
         $this->orderLoader                 = $orderLoader;
         $this->customerSession             = $customerSession;
         $this->customerRepository          = $customerRepository;
+        $this->customerLogin               = $customerLogin;
         $this->suiteLogger                 = $suiteLogger;
     }
 
     public function execute()
     {
+        $orderId = null;
         try {
             $quoteIdEncrypted = $this->getRequest()->getParam("quoteId");
             $quoteIdFromParams = $this->cryptAndCode->decodeAndDecrypt($quoteIdEncrypted);
             $quote = $this->quoteRepository->get((int)$quoteIdFromParams);
             $order = $this->orderLoader->loadOrderFromQuote($quote);
-            $customerId = $order->getCustomerId();
-            if ($customerId != null) {
-                $this->logInCustomer($customerId);
-            }
             $orderId = (int)$order->getId();
+            $customerId = $order->getCustomerId();
+
+            if ($customerId != null) {
+                $this->customerLogin->logInCustomer($customerId);
+            }
 
             $payment = $order->getPayment();
 
@@ -135,6 +144,7 @@ class Callback3Dv2 extends Action implements CsrfAwareActionInterface
             $data->setVendorName($this->config->getVendorname());
             $data->setMode($this->config->getMode());
             $data->setPaymentAction($this->config->getSagepayPaymentAction());
+            $data->setSaveToken((bool)$this->getRequest()->getParam("saveToken"));
 
             $this->requester->setRequestData($data);
 
@@ -149,12 +159,12 @@ class Callback3Dv2 extends Action implements CsrfAwareActionInterface
                 $this->javascriptRedirect('checkout/cart');
             }
         } catch (ApiException $apiException) {
-            $this->recoverCart->setShouldCancelOrder(true)->execute();
+            $this->recoverCart->setShouldCancelOrder(true)->setOrderId($orderId)->execute();
             $this->logger->critical($apiException);
             $this->messageManager->addError($apiException->getUserMessage());
             $this->javascriptRedirect('checkout/cart');
         } catch (\Exception $e) {
-            $this->recoverCart->setShouldCancelOrder(true)->execute();
+            $this->recoverCart->setShouldCancelOrder(true)->setOrderId($orderId)->execute();
             $this->logger->critical($e);
             $this->messageManager->addError(__("Something went wrong: %1", $e->getMessage()));
             $this->javascriptRedirect('checkout/cart');