+++ Model/PiRequestManagement.php
@@ -53,6 +53,8 @@ class PiRequestManagement implements \Ebizmarts\SagePaySuite\Api\PiManagementInt
         $data->setCcExpYear($requestData->getCcExpYear());
         $data->setCcLastFour($requestData->getCcLastFour());
         $data->setCcType($requestData->getCcType());
+        $data->setSaveToken($requestData->getSaveToken());
+        $data->setReusableToken($requestData->getReusableToken());
 
         $data->setJavascriptEnabled($requestData->getJavascriptEnabled());
         $data->setLanguage($requestData->getLanguage());