+++ Model/Token.php
@@ -6,7 +6,14 @@
 
 namespace Ebizmarts\SagePaySuite\Model;
 
+use Ebizmarts\SagePaySuite\Model\Api\Post;
 use Ebizmarts\SagePaySuite\Model\Logger\Logger;
+use Ebizmarts\SagePaySuite\Plugin\DeleteTokenFromSagePay;
+use Magento\Framework\Data\Collection\AbstractDb;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Model\Context;
+use Magento\Framework\Model\ResourceModel\AbstractResource;
+use Magento\Framework\Registry;
 
 /**
  * Sage Pay Token class
@@ -15,7 +22,7 @@ class Token extends \Magento\Framework\Model\AbstractModel
 {
 
     /**
-     * @var \Ebizmarts\SagePaySuite\Model\Api\Post
+     * @var Post
      */
     private $_postApi;
 
@@ -25,28 +32,43 @@ class Token extends \Magento\Framework\Model\AbstractModel
     private $_config;
 
     /**
-     * @param \Magento\Framework\Model\Context $context
-     * @param \Magento\Framework\Registry $registry
-     * @param \Magento\Framework\Model\ResourceModel\AbstractResource|null $resource
-     * @param \Magento\Framework\Data\Collection\AbstractDb|null $resourceCollection
+     * @var DeleteTokenFromSagePay
+     */
+    private $deleteTokenFromSagePay;
+
+    /**
+     * @var Logger
+     */
+    private $_suiteLogger;
+
+    /**
+     * @param Context $context
+     * @param Registry $registry
+     * @param Logger $suiteLogger
+     * @param Api\Post $postApi
+     * @param Config $config
+     * @param DeleteTokenFromSagePay $deleteTokenFromSagePay
+     * @param AbstractResource|null $resource
+     * @param AbstractDb|null $resourceCollection
      * @param array $data
      */
     public function __construct(
-        \Magento\Framework\Model\Context $context,
-        \Magento\Framework\Registry $registry,
+        Context $context,
+        Registry $registry,
         Logger $suiteLogger,
-        \Ebizmarts\SagePaySuite\Model\Api\Post $postApi,
+        Post $postApi,
         Config $config,
-        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
-        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
+        DeleteTokenFromSagePay $deleteTokenFromSagePay,
+        AbstractResource $resource = null,
+        AbstractDb $resourceCollection = null,
         array $data = []
     ) {
-    
         parent::__construct($context, $registry, $resource, $resourceCollection, $data);
-        $this->_suiteLogger = $suiteLogger;
-        $this->_logger = $context->getLogger();
-        $this->_postApi = $postApi;
-        $this->_config = $config;
+        $this->_suiteLogger           = $suiteLogger;
+        $this->_logger                = $context->getLogger();
+        $this->_postApi               = $postApi;
+        $this->_config                = $config;
+        $this->deleteTokenFromSagePay = $deleteTokenFromSagePay;
     }
 
     /**
@@ -75,7 +97,6 @@ class Token extends \Magento\Framework\Model\AbstractModel
      */
     public function saveToken($customerId, $token, $ccType, $ccLast4, $ccExpMonth, $ccExpYear, $vendorname)
     {
-
         if (empty($customerId)) {
             return $this;
         }
@@ -110,55 +131,33 @@ class Token extends \Magento\Framework\Model\AbstractModel
     }
 
     /**
-     * Delete token from db and Sage Pay
+     * @param $customerId
+     * @param $vendorname
+     * @return array
      */
-    public function deleteToken()
+    public function getCustomerTokensToShowOnAccount($customerId, $vendorname)
     {
-
-        //delete from sagepay
-        $this->_deleteFromSagePay();
-
-        if ($this->getId()) {
-            $this->delete();
+        $tokens = [];
+        $serverTokens = $this->getCustomerTokens($customerId, $vendorname);
+        foreach ($serverTokens as $token) {
+            $token['isVault'] = false;
+            $tokens[] = $token;
         }
+        return $tokens;
     }
 
     /**
-     * delete token using Sage Pay API
+     * Delete token from db and Sage Pay
+     * @throws NoSuchEntityException
+     * @throws \Exception
      */
-    private function _deleteFromSagePay()
+    public function deleteToken()
     {
-        try {
-            if (empty($this->getVendorname()) || empty($this->getToken())) {
-                //missing data to proceed
-                return;
-            }
+        //delete from sagepay
+        $this->deleteTokenFromSagePay->deleteFromSagePay($this->getToken());
 
-            //generate delete POST request
-            $data = [];
-            $data["VPSProtocol"] = $this->_config->getVPSProtocol();
-            $data["TxType"] = "REMOVETOKEN";
-            $data["Vendor"] = $this->getVendorname();
-            $data["Token"] = $this->getToken();
-
-            //send POST to Sage Pay
-            $this->_postApi->sendPost(
-                $data,
-                $this->_getRemoveServiceURL(),
-                ["OK"]
-            );
-        } catch (\Exception $e) {
-            $this->_logger->critical($e);
-            //we do not show any error message to frontend
-        }
-    }
-
-    private function _getRemoveServiceURL()
-    {
-        if ($this->_config->getMode() == Config::MODE_LIVE) {
-            return Config::URL_TOKEN_POST_REMOVE_LIVE;
-        } else {
-            return Config::URL_TOKEN_POST_REMOVE_TEST;
+        if ($this->getId()) {
+            $this->delete();
         }
     }
 
@@ -198,10 +197,17 @@ class Token extends \Magento\Framework\Model\AbstractModel
      */
     public function isOwnedByCustomer($customerId)
     {
-        if (empty($customerId) || empty($this->getId())) {
+        try {
+            if (empty($customerId) || empty($this->getId())) {
+                throw new NoSuchEntityException(
+                    __('Unable to delete token from Opayo: missing data to proceed')
+                );
+            }
+            return $this->getResource()->isTokenOwnedByCustomer($customerId, $this->getId());
+        } catch (NoSuchEntityException $e) {
+            $this->_suiteLogger->logException($e);
             return false;
         }
-        return $this->getResource()->isTokenOwnedByCustomer($customerId, $this->getId());
     }
 
     /**